+++ app/src/main/res/layout/custom_pref_top.xml
@@ -9,7 +9,6 @@
     android:paddingLeft="8dp"
     android:layout_marginLeft="6dp"
     android:layout_marginRight="6dp"
-    android:layout_marginTop="12dp"
     android:paddingEnd="?android:attr/scrollbarSize"
     android:paddingRight="?android:attr/scrollbarSize">
