--- app/src/main/res/xml/advanced_preferences.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <PreferenceCategory android:title="Advanced" android:layout="@layout/custom_pref_category">
-        <Preference
-            android:layout="@layout/custom_pref_whole"
-            android:icon="@drawable/bluetooth"
-            android:key="bluetooth_file"
-            android:title="View Bluetooth Server File"/>
-    </PreferenceCategory>
-
-</PreferenceScreen>
