--- app/src/main/res/xml/privacy_preferences.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <PreferenceCategory android:title="Privacy" android:layout="@layout/custom_pref_category">
-        <SwitchPreferenceCompat
-            android:title="Share Data"
-            android:layout="@layout/custom_pref_whole"
-            android:summary="Please enable to share data with the Treehouses Remote Team. This will help us to improve our services to give you the best possible experience!"
-            android:key="send_log"
-            />
-
-    </PreferenceCategory>
-
-</PreferenceScreen>
